Update README.md
Requirement Analysis is a crucial phase in the Software Development Life Cycle (SDLC) that involves identifying, gathering, analyzing, and documenting the needs and expectations of stakeholders for a software system. This process ensures that developers, clients, users, and all other stakeholders have a clear and shared understanding of what the software should do, how it should behave, and the constraints it must operate within.

Importance of Requirement Analysis
 It helps in understanding what the stakeholders expect from the software, reducing ambiguity.
A clearly defines the scope of the project, which helps in preventing scope creep. 
It provides a solid foundation for designing and developing the system.
It facilitates accurate estimation of project cost, resources, and time and also ensures that the final product meets the specified requirements, leading to higher customer satisfaction.